Update route list UI to pull from new API route and remove manual filtering.
Had to change the route schema so that day is a number rather than a string.

Weekday is Sunday - Saturday; 0 - 6

Moved the filtering, and sorting logic to the server under the new day call. Established changable timezones so the server can be any timezone we want.

Note that the client will not really work at this time until the API call in the client is updated along with the new logic.